Slightly refactor Class2D and RIRClass2D
